Implemented 'class' slot allocation.
This one is sort of tricky. Memory for these slots should be allocated
only once per whole lineage of classes. I wanted to carry on with that
'closed vector' approach, so we have two cases: direct class-slots and
inherited class-slots. The first ones should get their fresh new chunk
of memory, and accessors of the second group should be copied from one
of the superclasses: in this way they retain access to the data vector
with shared slot values.

Also the slot groups are removed from `install-direct-accessors!` args.
They aren't helpful because they are just <slots>, not <effective-slots>.
Instead I sweep though the inheritance tree directly to get what I need.

The code is in kinda messy state right now, should be cleaned up after
the virtual slots are ready.